added some files to setup refactor of the patient component
