+++ project/web/app/biometric-monitor/PatientComponent/AddPatientForm.tsx
@@ -1,0 +1,6 @@
+
+"use client"
+
+export default function AddPatientForm(){
+
+}
